Add custom icons to Dashbaord and Login forms
Introduced a `ComponentResourceManager` in `Dashbaord.Designer.cs`
and `Login.Designer.cs` to manage resources. Set the `Icon`
property for both forms to display a custom icon. Updated
`Dashbaord.resx` and `Login.resx` to include the icon data in
Base64 format and referenced the `System.Drawing` assembly.

Added new `taskbaricon.ico` and `taskbaricon.png` files for
application branding. Adjusted layout and initialization of
controls in the forms to accommodate the new icon and resource
manager. These changes improve the application's visual
appearance and branding.